Fix coroutine bug
At the beginning or end of certain coroutines, it would stall
	until a user performed some action.
This was because there were no events to pull.
This is fixed by adding `os.startTimer(0)`
	once per Subscribe/UnSubscribe and once per run.

There was also a bug when an event was unsubscribed.
This was due to removing the coroutine itself from the list,
	and then trying to call coroutine.status(nil)
This was solved by adding checks for `not v.co`